pyproject.toml
[project]
name = "project-a"
version = "0.1.0"
description = "Streamlit image search apps"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "sentence-transformers>=2.2.2",
  "Pillow>=10.0.0",
  "numpy>=1.24.0",
  "pandas>=2.0.0",
  "plotly>=5.15.0",
  "pathlib2>=2.3.7",
  "typing-extensions>=4.7.0",
  "torch>=2.8.0",
  "streamlit>=1.50.0",
  "kaggle>=1.7.4.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E","F","I","UP"]
ignore = []

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true

[tool.hatch.build.targets.wheel]
packages = ["project_a"]   # if you later add a package dir; otherwise remove this

# If you also build sdists and want data included:
[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    # ... other includes like README, pyproject.toml, etc.
]
